# perf: fix perf tests to support generic compat pmu