--- a/perf/perf_genericevents.py.data/raw_code.cfg
+++ b/perf/perf_genericevents.py.data/raw_code.cfg
@@ -1,3 +1,17 @@
+[GENERIC_COMPAT_PMU]
+cpu-cycles = 0x600f4
+instructions = 0x500fa
+stalled-cycles-frontend = 0x100f8
+branch-misses = 0x400f6
+cache-misses = 0x400f0
+L1-dcache-load-misses = 0x400f0
+L1-dcache-store-misses = 0x300f0
+L1-icache-load-misses = 0x200fc
+LLC-load-misses = 0x300fe
+branch-load-misses = 0x400f6
+dTLB-load-misses = 0x300fc
+iTLB-load-misses = 0x400fc
+
 [POWER8]
 cpu-cycles = 0x1e
 stalled-cycles-frontend = 0x100f8